Вывод популярных фильмов по жанру и годам - исправление ошибок

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -35,7 +35,7 @@ public class FilmService {
         return filmDbRepository.getAll();
     }
 
-    public Film getFilmById(Long id) {
+    public Film getFilmById(Integer id) {
         return filmDbRepository.getById(id).orElseThrow(() -> new NotFoundException("Ошибка! Фильма с заданным " +
                 "идентификатором не существует"));
     }
